別プロジェクトに2種類の実装を分離(Minoクラスを実装したものと、未実装のもの) 任意のブロック数Nにおけるミノの全パターンの出力を実装 主要な処理をメソッド化 繰り返しを省略
