+++ CreatingMino.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 17.7.34009.444
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CreatingMino", "CreatingMino\CreatingMino.csproj", "{9494E9C9-EB27-4C26-8F64-9D1F1171EBE5}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OldCreatingMino", "OldCreatingMino\OldCreatingMino.csproj", "{07E7F641-355A-4B29-91EA-46A3FC900909}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +17,10 @@ Global
 		{9494E9C9-EB27-4C26-8F64-9D1F1171EBE5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9494E9C9-EB27-4C26-8F64-9D1F1171EBE5}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{9494E9C9-EB27-4C26-8F64-9D1F1171EBE5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{07E7F641-355A-4B29-91EA-46A3FC900909}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{07E7F641-355A-4B29-91EA-46A3FC900909}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{07E7F641-355A-4B29-91EA-46A3FC900909}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{07E7F641-355A-4B29-91EA-46A3FC900909}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
